requirements.txt
ipython==8.28.0
sagemath_standard==10.4
